dev_requirements.txt
pytest-dbt-adapter>=0.4.0
sqlalchemy>=1.3.23,<2.0
prospector==1.3.1